wayland: Resize wl_egl_window from the correct thread with an active context
According to Nvidia documentation [1] wl_egl_window_resize should only
be called when the context is current and from the correct thread.

This refactors the code to cache a thread-safe buffer size when
the geometry changes from the main thread but defer the actual resize
call.

On the next frame we:
 - potentially create/delete an egl surface as needed
 - call eglMakeCurrent
 - resize the egl window

This patch also re-enables threaded rendering for Nvidia now the root
bug has been fixed.

[1] https://github.com/NVIDIA/egl-wayland2

Fixes: QTBUG-95817
Change-Id: Ic987880c3d4a1220044fb4e8b3eeac0a4e8c06ea
Reviewed-by: David Redondo <qt@david-redondo.de>
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>