Sync rhi backingstore d3d12 behavior with Quick
When adding the D3D12 code path in widgets it seems it was ignored that
the retry-with-WARP logic was present both for D3D11 and 12 in Qt Quick.

For consistency, the same logic must be used. Right now widgets only has
the retry logic for D3D11.

Change-Id: I08d5728acfb9a68db7e0a4b0477050f08546ffac
Reviewed-by: Christian Strømme <christian.stromme@qt.io>
(cherry picked from commit 54816a8f9af706d9a14cd2501fcaf74962bbdbd8)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>